Re-enable metadce main module test
This was disabled back #9120 because it was deemed to fragile.
This change re-enables it but using a C program rathe than
a C++ one so it doesn't include the C++ standard library which
makes it way smaller and without a gazillion obscure exports.

Also switch to use `self.build` to compile these tests so that
we get the C vs C++ handling that exists here.

Tracking the symbol list used during a minimal MAIN_MODULE=1
build is useful for my upcoming PIC ABI changes to see the
effect it has.