Set RELOCATABLE if `-fPIC` is passed on the command line
Also avoid adding `-fPIC` a second time if it already exists.

This is a partial fix for #14488 since now building with `-fPIC -s
USE_LIBJPEG=1` will correctly cause the `pic` version of libjpeg to be
built and installed.

This is hard to test for since it involves something that is mostly and
internal detail (which verion of libjpeg is build when copiling with `-c`).

I decided to instead test for another side effect of `-s RELOCATABLE`
which is the `-fvisibility=default` compiler flag.