Make llvm version check into an error
Using emscripten with the wrong llvm version is not possible and we
should be clear about that.  The fact that it has historically just
been an error has led to folks trying to use mismtached versions and
reporting bugs about it.

Trying to make emscripten work with an incompatible version of llvm will
most likely require a patch and so a modified expected version can be
part of any such patch.