Remove incorrect/non-working (and used?) implementation of pthread_kill
Our implementation of `pthread_kill` seems to be based on the assumption
that this function is supposed to kill and given thread.  This is not
true.  Instead `pthread_kill` works more like `kill` in that is delivers
and async signal to a given thread.  There are many such signals
available such as SIGUSR1, SIGUSR2, SIGKILL, and SIGTERM.   All that
pthread_kill does that is different from `kill` is that it enables the
caller to decided which thread the signal handler should run in.  There
is no signal that can sent that reults in a termination of a given
thread.  Sending SIGKILL (which is unblockable) will still bring down
the entire program.  There is no pthread API for asynronously killing a
single thread, only pthread_cancel, which requires some cooperation from
the thread being terminated.

The currently implementation of `pthread_kill` is based on
`worker.terminate()`, which will kill the worker even if it is in a busy
loop (how does this work at the OS level I wonder?), is problematic
becuse the thread has no way to clean up or free resources which can
lead to memory leaks.

Finally, `pthread_kill` does not currently work under chrome (brings
down the entire tab) so this leads me to believe that it is very
unlikely anyone is relying on the feature.

Fixes: #14872