Remove assembly vesion of emscripten_is_main_browser_thread
This function was originally implemented as global.get because like
`emscripten_is_main_runtime_thread` because it was called frequently
from native threading primitives.  However, recent changes have all but
removed this function from the codebase so there is no need to cache
this in a wasm global anymore.

The only two remaining users of this function are:
system/lib/fetch/asmfs.cpp
system/lib/fetch/emscripten_fetch.cpp

The former is deprecated and later is not performance critical.