Add #! line to top of JS file
Previously we could only add this when we detect that we running under
autoconf.

With this change we now default to outputting a JS file that is
executable unless as long as we are possibly targetting node
(ENVIRONMENT_MAY_BE_NODE) and we are not optimizing for size (`-Oz` or
`-Os`).

Also, conditionally add `--experimental-wasm-threads` to the node
command (similar to what we did in #15590) used in the #! line.