tests: Add --skip-standalone option to skip over standalone wasm engine tests. NFC
This is useful for users to doing have a standalone wasm engine
(e.g. wasmtime or wasmer) installed.

also avoids silently skipping these tests if the standalone engine is
not configured currently (for example on the bots) but forcing the
caller to opt out explicitly.

Also, add a way to more explicitly skip standalone engines for a given
test.