Run textwrap.dedent on text files created by the test runner
This means we can indent out test snippets in the python code and should
never need to use column 1.

The files on disc that the test runner writes will have sane/normal
indentation even if we use extra indentation in the python source to
make things readable.

For example:

```
    create_file('a.c', '''
       int main()
         printf("hello\n");
       }
    ''')
```

Would generate a file on disc containing:

```
int main()
  printf("hello\n");
}
```