Avoid using RELOCATABLE by default when building with MAIN_MODULE
By building main module as non-relocatable we can improve the code size
and performance of the main module.  There is no good reason to build
the main module as relocatable, only the side module really need this.

One downside is that this reduces our coverage of the `RELOCATABLE`
option to just a few isolated test cases. I wonder if we should consider
removing this option unless there are other compelling use cases?

This change depends and llvm-side change: https://reviews.llvm.org/D91577