Strip the name section when including dwarf information
To enable this we separate the logic for stripping of dwarf information
and the names section.

The preservation of the names section occurs only if the interal
EMIT_NAME_SECTION is (which only happens if `--profiling-funcs` is set).
In addition create a new `-gnames` flag which is essentially an alias of
the less-well-named `--profiling-funcs`.

The effect of this change is that normal dward debug builds will no
longer contain a names section.  This seems to be desirable for most
users since the name section can be very large and is redundant in
face of dwarf info.  User can opt into both by specifying `-g` in
addition of `--profiling-funcs`.