Attempt to terminate workers cooperatively
This change is designed fix address pthread thread flakiness around
process termination.  For example see #19683 and #15014.

In some cases I believe that workers are being terminated while they
have postMessage messages in their queue, so messages are getting lost.

In other cases we are seeing message arrive after the call to
`terminate()`.

This change attempts to avoid using `terminate()` at all and attempts
to cooperatively shut down the worker.  We only use `terminate()` if the
thread doesn't respond in a certain amount of time.