Allow certain command line settings to be additive
For certain command line settings (specifically most settings that take
a list) it is useful to be able to add to the existing setting rather
than replace it.

For example, the `-sEXPORTED_FUNCTIONS` argument.  In a build system
where different components can inject link flags it makes sense that
each component be able add to the list of exported functions without
clobbering the current list.

We have had this feature requested several times in the past. For
example, from the bazel toolchain folks.