Put in place some minimum browser versions
These are the values we use when `-sLEGACY_VM_SUPPORT` is set or when
the users set one of the `-sMIN_XX_VERSION` settings to zero.  Zero has
always implied "oldest supported version".

The upside of this is that when babel is run now with
`-sLEGACY_VM_SUPPORT` we get a babel config file that looks like this:

```
{
  "sourceType": "script",
  "targets": {
    "chrome": "45",
    "firefox": "34",
    "safari": "9.0.0",
    "node": "10.19.0"
  }
}
```

Having these minimums in place will allows us to remove support for
older browsers from the codebase over time.  For example, as a followup
to this change we should be able to remove out polyfills for Promise and
for Object.assign.