Fix closure warnings in JS library code
For some reason closure doesn't report these real issues in the current
configuration but they show up in the `MODULARIZE_INSTANCE`
configuration that I'm working on.

Most of these issues consist of adding default parameters to functions
declared on objects such as `FS`.

In addition there are a couple of other changes that are all real fixes:

- ErrnoError unified between old FS and wasm FS.  This should have been
  done as part of #21149
- Default initializer for `canvasResizedCallbackTargetThread` removed
  since `JSEvents.getTargetThreadForEventCallback()` will always return
  undefined when called with no arguments
- Extra arguments to copyIndexedColorData removed in library_sdl.js
  since it only ever called with one arg.
- Sorting callback for JSEvents.deferredCalls.sort updated to correctly
  return number rather than a boolean.

This change is basically code size neutral with a few tests coming in
a byte or two heavier and others a byte or two lighter.