Remove USE_ES6_IMPORT_META setting
This setting was added (IIUC) as a temporary workaround for the lack or
support for `import.meta` in browser and tooling.

However `import.meta` has been at stage 4 for almost 5 years now:
https://github.com/tc39/proposal-import-meta/pull/21

Webback (released in 2020) also has builtin support for `import.meta`,
so I would hope the use case for disabling this setting no longer
exists.

See #9234 and #8729