Don't change getUnsharedTextDecoderView behaviour based on -Oz. NFC
Core semantics should not change with `-Oz`.

I also widened the heursitc so we assume shared for all `HEAPXX` object
which should mitigate any size regression here.

There is only a single caller of `getUnsharedTextDecoderView` that
doesn't pass a `HEAP` object: UTF8ArrayToString.  So this change should
have very little code size impact.

If we want to claw this regression back we could look into a version of
`UTF8ArrayToString` that only operates on heaps.